Add phone formatter
In this file, the data from the csv file is consulted, where the
telephone numbers are requested and stored. Some phone numbers are not
consistent with the desired or correct format. The phones are corrected
by being passed through a method that includes if/elsif/else clauses
that will evaluate the state of the object and modify it if necessary.
At the end, the phone numbers are printed with associated data from the
assistant.