Add registration date parser
In this file the values associated with the reg_date column of the csv
file will be captured. With these strings, instances of the Time class
will be created and stored as an array of dates. With custom methods you
obtain values such as frequent registration times and the day of the
week with the most registration frequency. In the end it is printed as
a string when it is appropriate to disclose information on different
platforms.